Разбивает файл main на модули
